Simplying Activity bar, MixedTree will use most stuff from SFMLTree
Changed the settings names
